backend/requirements.txt
FLASK
Flask-SQLAlchemy
Flask-Cors
python-dotenv